4425 - GreaterThan - add test cases
This test case with `<10, 9>` is different from the `<111, 11>` in the way that it has arguments of different lengths AND lower number has greater digits. The [solution I've recently posted to this problem](https://github.com/type-challenges/type-challenges/issues/21721) appeared to have this bug and it had been figured out only thanks to @Yueleng.

Also, I've added a test case with a big number (within save Number range) that I've notes some other solutions are failing on